Fixed a bug in VideoMaterial where it would previously try to play in the constructor, even if no file was set. If a file was defined at a later time, it would not play, because the net connection was already initialized.
This meant that the below would not work, because once file was set, the netstream had already tried playing, and was gonna fail.

videoMat = new VideoMaterial();
videoMat.file = 'myfile.flv';

By the time the stream failed, the file had been defined, and the error was reported incorrectly as 'could not find myfile.flv' even though it's there.


git-svn-id: https://away3d.googlecode.com/svn/trunk/fp10/Away3D@2534 6da1ee2c-932f-0410-8bb4-73b2b089c45a